wip - Migrate string types/constraints out of validation package
