+++ packages/overture-schema-divisions-theme/src/overture/schema/divisions/division_boundary/models.py
@@ -12,7 +12,6 @@ from overture.schema.core.ref import Reference, Relationship
 from overture.schema.core.types import (
     CountryCode,
     Id,
-    RegionCode,
 )
 from overture.schema.core.validation import (
     exactly_one_of,
@@ -24,6 +23,7 @@ from overture.schema.foundation.primitive import (
     GeometryType,
     GeometryTypeConstraint,
 )
+from overture.schema.foundation.string import RegionCode
 
 from ..division import Division
 from ..enums import PlaceType
