+++ packages/overture-schema-divisions-theme/src/overture/schema/divisions/models.py
@@ -3,8 +3,9 @@ from typing import Annotated, NewType
 from pydantic import ConfigDict, Field
 
 from overture.schema.core import StrictBaseModel
-from overture.schema.core.types import Id, StrippedString
+from overture.schema.core.types import Id
 from overture.schema.divisions.enums import PlaceType
+from overture.schema.foundation.string import StrippedString
 
 DivisionId = NewType(
     "DivisionId", Annotated[Id, Field(min_length=1, description="ID of the division")]
