+++ packages/overture-schema-foundation/src/overture/schema/foundation/constraint/__init__.py
@@ -11,7 +11,9 @@ Modules
 constraint : module
     Base constraint type.
 collection : module
-    Constraints for collections
+    Constraints for collection fields.
+string : module
+    Constraints for string fields.
 
 """
 
@@ -20,9 +22,33 @@ from .collection import (
     UniqueItemsConstraint,
 )
 from .constraint import Constraint
+from .string import (
+    CountryCodeConstraint,
+    HexColorConstraint,
+    JsonPointerConstraint,
+    LanguageTagConstraint,
+    NoWhitespaceConstraint,
+    PatternConstraint,
+    PhoneNumberConstraint,
+    RegionCodeConstraint,
+    StringConstraint,
+    StrippedConstraint,
+    WikidataIdConstraint,
+)
 
 __all__ = [
     "Constraint",
     "CollectionConstraint",
     "UniqueItemsConstraint",
+    "CountryCodeConstraint",
+    "HexColorConstraint",
+    "JsonPointerConstraint",
+    "LanguageTagConstraint",
+    "NoWhitespaceConstraint",
+    "PatternConstraint",
+    "PhoneNumberConstraint",
+    "RegionCodeConstraint",
+    "StringConstraint",
+    "StrippedConstraint",
+    "WikidataIdConstraint",
 ]
