+++ packages/overture-schema-foundation/src/overture/schema/foundation/string.py
@@ -1,0 +1,152 @@
+from typing import Annotated, NewType
+
+from pydantic import Field
+
+from overture.schema.foundation.constraint import (
+    CountryCodeConstraint,
+    HexColorConstraint,
+    JsonPointerConstraint,
+    LanguageTagConstraint,
+    NoWhitespaceConstraint,
+    PhoneNumberConstraint,
+    RegionCodeConstraint,
+    StrippedConstraint,
+    WikidataIdConstraint,
+)
+
+HexColor = NewType(
+    "HexColor",
+    Annotated[
+        str,
+        HexColorConstraint(),
+        Field(
+            description="A color represented as an #RRGGBB or #RGB hexadecimal string, for example #ff0000 for pure red."
+        ),
+    ],
+)
+HexColor.__doc__ = """
+HexColor : NewType
+    A color represented as an #RRGGBB or #RGB hexadecimal string.
+
+    For example:
+
+    - "#ff0000" for pure red 🟥
+    - "#ffa500" for bright orange 🟧
+    - "#000000" or "#000" for black ⬛
+"""
+
+
+JsonPointer = NewType(
+    "JsonPointer",
+    Annotated[
+        str,
+        JsonPointerConstraint(),
+        Field(description="A JSON Pointer (as described in RFC-6901)"),
+    ],
+)
+JsonPointer.__doc__ = """
+JsonPointer : NewType
+    A JSON Pointer
+
+    As described in `the JSON Pointer specification, RFC-6901 <https://rfc-editor.org/rfc/rfc6901.html>`
+"""
+
+
+LanguageTag = NewType(
+    "LanguageTag",
+    Annotated[
+        str,
+        LanguageTagConstraint(),
+        Field(
+            description="A BCP-47 language tag",
+        ),
+    ],
+)
+LanguageTag.__doc__ = """
+LanguageTag : NewType
+    A BCP-47 language tag.
+
+    As described in `Tags for Identifying Languages, BCP-47 <https://www.rfc-editor.org/rfc/bcp/bcp47.txt>`
+"""
+
+
+StrippedString = NewType(
+    "StrippedString",
+    Annotated[
+        str,
+        StrippedConstraint(),
+        Field(
+            description="A string without leading or trailing whitespace",
+        ),
+    ],
+)
+StrippedString.__doc__ = """
+StrippedString : NewType
+    A string without leading or trailing whitespace.
+"""
+
+
+NoWhitespaceString = NewType(
+    "NoWhitespaceString",
+    Annotated[
+        str,
+        NoWhitespaceConstraint(),
+        Field(description="A string that contains no whitespace characters"),
+    ],
+)
+NoWhitespaceString.__doc__ = """
+NoWhitespaceString : NewType
+    A string that contains no whitespace characters.
+"""
+
+
+CountryCode = NewType(
+    "CountryCode",
+    Annotated[
+        str,
+        CountryCodeConstraint(),
+        Field(description="An ISO 3166-1 alpha-2 country code"),
+    ],
+)
+CountryCode.__doc__ = """
+CountryCode : NewType
+    An ISO-316601 alpha-2 country code.
+"""
+
+RegionCode = NewType(
+    "RegionCode",
+    Annotated[
+        str,
+        RegionCodeConstraint(),
+        Field(description="An ISO 3166-2 principal subdivision code"),
+    ],
+)
+RegionCode.__doc__ = """
+RegionCode : NewType
+    An ISO 3166-2 principal subdivision code.
+"""
+
+WikidataId = NewType(
+    "WikidataId",
+    Annotated[
+        str,
+        WikidataIdConstraint(),
+        Field(description="A wikidata ID, as found on https://www.wikidata.org/"),
+    ],
+)
+WikidataId.__doc__ = """
+WikidataId : NewType
+    A wikidata ID, as found on https://www.wikidata.org/.
+"""
+
+
+PhoneNumber = NewType(
+    "PhoneNumber",
+    Annotated[
+        str, PhoneNumberConstraint(), Field(description="An international phone number")
+    ],
+)
+PhoneNumber.__doc__ = """
+PhoneNumber : NewType
+    An international telephone number.
+"""
